can add estimates on a new issue in decimal

+++ src/timepiece/forms.py
@@ -786,7 +786,7 @@ class IssueStatusForm(forms.ModelForm):
 
 class IssueForm(forms.ModelForm):
 
-    estimated_hours = forms.IntegerField(required=False, initial=1)
+    estimated_hours = forms.FloatField(required=False, initial=1)
 
     class Meta:
         model = timepiece.Issue
